more tests and refactoring
