--- peakipy/cli/edit_fits_app/templates/index.html
@@ -1,16 +1,0 @@
-{% extends base %}
-
-{% block title %}Bokeh Crossfilter Example{% endblock %}
-
-{% block preamble %}
-<style>
-  @import url(https://fonts.googleapis.com/css?family=Noto+Sans);
-  body {
-    font-family: 'Noto Sans', sans-serif;
-    -webkit-font-smoothing: antialiased;
-    text-rendering: optimizeLegibility;
-    color: #fff;
-    background: #2F2F2F;
-  }
-</style>
-{% endblock %}
