--- peakipy/cli/edit_fits_app/theme.yaml
@@ -1,18 +1,0 @@
-attrs:
-    Figure:
-        background_fill_color: 'white'
-        border_fill_color: '#2F2F2F'
-        outline_line_color: '#444444'
-
-    Axis:
-        axis_line_color: "white"
-        axis_label_text_color: "white"
-        major_label_text_color: "white"
-        major_tick_line_color: "white"
-        minor_tick_line_color: "white"
-        minor_tick_line_color: "white"
-    Grid:
-        grid_line_dash: [6, 4]
-        grid_line_alpha: .3
-
-text_color: "white"
